Classes moved to Entities file
